CU-863h3ztv4: Quartz job to check eveara status

+++ newm-server/src/main/kotlin/io/newm/server/features/distribution/model/FeaturedArtist.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class FeaturedArtist(
     @SerialName("artist_id")
-    val artistId: Int,
+    val artistId: Long,
     @SerialName("name")
     val name: String
 )
